+++ packages/granularity/src/components/GrCollapse/config.ts
@@ -1,6 +1,9 @@
 import { defineGranularComponent } from '@feugene/unocss-preset-granular/contract'
 
+import { grCollapseSafelist } from './safelist'
+
 export const grCollapseConfig = defineGranularComponent(import.meta.url, {
   name: 'GrCollapse',
   dependencies: ['GrCard', 'GrIcon'],
+  safelist: grCollapseSafelist,
 })